requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pyodbc>=4.0.39
python-dotenv>=1.0.0
google-generativeai>=0.2.0
requests>=2.31.0
pydantic>=2.4.2
plotly>=5.17.0
sqlalchemy>=2.0.23
pandas>=2.1.2
numpy>=1.26.1
pytest>=7.4.3
black>=23.10.1
isort>=5.12.0
python-multipart>=0.0.6